backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
langchain
langchain-community
langchain-text-splitters
sentence-transformers
python-pptx
python-docx
pandas
openai
pypdf
python-dotenv
openpyxl
